airflow-local/requirements.txt
apache-airflow==2.10.4
azure-storage-queue==12.11.0
azure-identity==1.19.0
